refactor(backend): change function declaration
